Moved types to domain folder
