+++ src/Kingdom/Overview/domain/addTerrainFeatures.ts
@@ -1,0 +1,5 @@
+import { Tile } from "./types";
+
+export const addTerrainFeatures = (tiles: Array<Tile>) => {
+
+};
