+++ src/Kingdom/Overview/domain/types.ts
@@ -1,0 +1,3 @@
+export type Kingdom = { name: String, terrain: Terrain };
+export type Terrain = { rowSize: number, columnSize: number, tiles: Array<Tile> };
+export type Tile = { id: string, x: number, y: number, type: string };
